+ checks for tricksurf gamemode
This was causing `mom_restart_stage` to break zones when doing a manual teleport.

+++ mp/src/game/shared/momentum/mom_system_tricks.cpp
@@ -615,11 +615,17 @@ void CTrickSystem::OnTrickFailed(CTrickAttempt *pAttempt)
 
 void CTrickSystem::ClearTrickAttempts()
 {
+    if (!g_pGameModeSystem->GameModeIs(GAMEMODE_TRICKSURF))
+        return;
+
     m_vecCurrentTrickAttempts.PurgeAndDeleteElements();
 }
 
 void CTrickSystem::PostPlayerManualTeleport(CMomentumPlayer *pPlayer)
 {
+    if (!g_pGameModeSystem->GameModeIs(GAMEMODE_TRICKSURF))
+        return;
+
     ClearTrickAttempts();
 
     const auto iTracked = m_iTrackedTrick;
